+++ rust/crates/imperialism-app/src/ui/offer_sheet.rs
@@ -441,7 +441,7 @@ mod tests {
 
     fn fixture_state() -> GameState {
         let selected_nation = peek_save_header(BEGINNING_OF_GAME)
-            .and_then(|header| NationId::try_new(header.active_nation))
+            .and_then(|header| NationId::try_new(usize::from(header.active_nation)))
             .unwrap_or(NationId::new(0));
         LegacySaveV62::parse(BEGINNING_OF_GAME).game_state(LegacyGameStateContext {
             crt_rand_state: 1,
